[thin_debug] Factor out reusable componments
